Some small changes which also made the image smaller!
I have made small changes and also managed to get the size down -- even with adding
dumb-init to the image! I have also merged the scripts to one script for both
backup and restore

I have also fixed the bug that caused cronbackups to not run due a small
oversight when I made mt changes since the backup.sh file is on the $PATH, oops.